refactor(tests): extract shared wait utilities from e2e for cross-suite use
- Create tests/shared/ with TypeScript config, barrel export, and ambient types
- Extract timing constants (E2E_TIMING) to tests/shared/timing-constants.ts
- Extract test logger (E2ELogger) to tests/shared/test-logger.ts as TestLogger
- Extract all 8 wait utility functions to tests/shared/wait-utilities.ts
- Update tests/e2e/helpers/ to re-export from shared (zero backward-compat breaks)
- Add tests/shared/wdio-types.d.ts for WebdriverIO ambient type support
- Update e2e tsconfig to include shared ambient declarations

Phase 1 verified: 125 e2e tests passing, 0 new integration failures introduced.
Enables Phase 2-4 integration test migration to import shared wait utilities.
