+++ tests/e2e/helpers/logger.ts
@@ -1,67 +1,11 @@
 /**
- * Valid logging levels for E2E tests.
+ * E2E Logger — re-exports TestLogger from shared as E2ELogger for backward compatibility.
+ *
+ * All existing imports of E2ELogger and LogLevel from this file continue to work unchanged.
  */
-export type LogLevel = 'info' | 'warn' | 'error' | 'debug';
 
-/**
- * Enhanced Logger for E2E tests.
- * Wraps console logging with timestamps and consistent formatting.
- * Only logs 'debug' level if E2E_DEBUG env var is set.
- */
-export class E2ELogger {
-    /**
-     * Logs an informational message.
-     * @param context The context/component name (e.g. 'menu_bar')
-     * @param message The message to log
-     * @param data Optional data object to log
-     */
-    static info(context: string, message: string, data?: any) {
-        this.log('info', context, message, data);
-    }
+// Re-export LogLevel type so existing `import { LogLevel } from './logger'` sites compile.
+export type { LogLevel } from '../../shared/test-logger';
 
-    /**
-     * Logs an error message.
-     * @param context The context/component name
-     * @param message The error message
-     * @param error The error object or data
-     */
-    static error(context: string, message: string, error?: any) {
-        this.log('error', context, message, error);
-    }
-
-    /**
-     * Logs a warning message.
-     * @param context The context/component name
-     * @param message The warning message
-     * @param data Optional data
-     */
-    static warn(context: string, message: string, data?: any) {
-        this.log('warn', context, message, data);
-    }
-    /**
-     * Logs a debug message. Only visible if process.env.E2E_DEBUG is set.
-     * @param context The context/component name
-     * @param message The debug message
-     * @param data Optional data
-     */
-    static debug(context: string, message: string, data?: any) {
-        if (process.env.E2E_DEBUG) {
-            this.log('debug', context, message, data);
-        }
-    }
-
-    /**
-     * Internal log handler.
-     * @private
-     */
-    private static log(level: LogLevel, context: string, message: string, data?: any) {
-        const timestamp = new Date().toISOString();
-        const prefix = `[${timestamp}] [${level.toUpperCase()}] [${context}]`;
-
-        if (data) {
-            console.log(`${prefix} ${message}`, data);
-        } else {
-            console.log(`${prefix} ${message}`);
-        }
-    }
-}
+// Re-export TestLogger under the E2ELogger name used throughout e2e tests.
+export { TestLogger as E2ELogger } from '../../shared/test-logger';
